requirements.txt
Django==2.0.2
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Markdown==2.6.11
Pillow==5.0.0
PyJWT==1.6.0
pytz==2018.3
